+++ packages/fuels-programs/src/assembly.rs
