+++ packages/fuels-programs/src/assembly/contract_call.rs
