+++ packages/fuels-programs/src/assembly/cursor.rs
